--- packages/celotool/src/cmds/generate/private-key-to-address.ts
@@ -1,23 +1,0 @@
-/* tslint:disable no-console */
-import { privateKeyToAddress } from 'src/lib/generate_utils'
-import * as yargs from 'yargs'
-
-interface AccountAddressArgv {
-  privateKey: string
-}
-
-export const command = 'private-key-to-address'
-
-export const describe = 'command for generating account address from private key'
-
-export const builder = (argv: yargs.Argv) => {
-  return argv.option('private-key', {
-    type: 'string',
-    description: 'private key',
-    demand: 'Please, specify an account private key',
-  })
-}
-
-export const handler = async (argv: AccountAddressArgv) => {
-  console.log(privateKeyToAddress(argv.privateKey))
-}
